Rename runPage -> runPageTooltips
